refactor: 테스트 api url 리팩토링
테스트 용 api 태그 분리
