+++ app/routers/api.py
@@ -1,7 +1,5 @@
 from fastapi import APIRouter
-from . import images
-from . import chats
-from . import voices
+from . import chats, voices, images, tests
 
 router = APIRouter(
     prefix="/api/v1"
@@ -10,3 +8,4 @@ router = APIRouter(
 router.include_router(chats.router)
 router.include_router(images.router)
 router.include_router(voices.router)
+router.include_router(tests.router)
